requirements.txt
PyYAML==5.4.1
tinydb==4.3.0